Moved Tests in a better place
